+++ src/Stormancer.Plugins/Steam/Stormancer.Server.Plugins.Steam/Changelog.rst
@@ -13,6 +13,7 @@ unreleased
 ----------
 - revert plugin to normal state
 - Fix queryUserIds (no error if a steamId is not found in the system)
+- Improve steam friends
 
 4.0.4
 -----